Move celery.py to toplevel rdwatch module
