--- rdwatch/scoring/__init__.py
@@ -1,3 +1,0 @@
-# placeholder for version number, will be replaced by
-# poetry-dynamic-versioning in production
-__version__ = '0.0.0dev0'
